feat: support direction-scoped stop time updates
 Add shared direction parsing and use it to validate, resolve, and build direction-specific stop-time-update topics while
  preserving existing unfiltered behavior.
